+++ src/features/celebrities/components/CelebrityCategorySelect/index.ts
@@ -1,0 +1,3 @@
+import CelebrityCategorySelect from './CelebrityCategorySelect';
+
+export default CelebrityCategorySelect;
